Support for dpy2, minor changes to the code itself.
