+++ chessgame/commands/scoreboard.py
@@ -5,15 +5,15 @@ import math
 from typing import Union
 
 import discord
+
 from redbot.core import commands
-from redbot.core.utils.menus import (DEFAULT_CONTROLS, menu,
-                                     start_adding_reactions)
+from redbot.core.utils.menus import DEFAULT_CONTROLS, menu, start_adding_reactions
 from redbot.core.utils.predicates import ReactionPredicate
 
 from .main import chess
 
 
-class ScoreboardCommands:
+class ScoreboardCommands():
     """Scoreboard related commands"""
 
     @chess.group(name="scoreboard")
