+++ chessgame/constants.py
